cargo_paralegal_flow
Struct ClapArgs

pub struct ClapArgs {
    pub result_path: PathBuf,
    pub relaxed: bool,
    pub strict: bool,
    pub target: Option<String>,
    pub abort_after_analysis: bool,
    pub attach_to_debugger: Option<Debugger>,
    pub anactrl: ClapAnalysisCtrl,
    pub marker_control: MarkerControl,
    pub dump: ParseableDumpArgs,
    pub build_config: Option<PathBuf>,
    pub cargo_args: Vec<String>,
}

Arguments as exposed on the command line.

You should then use try_into to convert this to [Args], the argument structure used internally.

Fields
result_path: PathBuf

Where to write the resulting GraphLocation (defaults to flow-graph.json)

relaxed: bool

Emit warnings instead of aborting the analysis on sanity checks

This is now the default behavior and this flag is deprecated. Use --strict to turn off this behavior.

strict: bool

Emit errors instead of warnings for potential soundness risks

target: Option<String>

Run paralegal only on this crate

abort_after_analysis: bool

Abort the compilation after finishing the analysis

attach_to_debugger: Option<Debugger>

Attach to a debugger before running the analyses

anactrl: ClapAnalysisCtrl

Additional arguments that control the flow analysis specifically

marker_control: MarkerControl

Additional arguments which control marker assignment and discovery

dump: ParseableDumpArgs

Additional arguments that control debug args specifically

build_config: Option<PathBuf>

Pass through for additional cargo arguments (like –features)

cargo_args: Vec<String>
Implementations

impl ClapArgs

pub fn hash_config(&self, hasher: &mut impl Hasher)

pub fn get_build_config(&self) -> Option<&Path>